Extract shared networking-enabled config check into libs/networking
The feature-flag + $VX_CONFIG_ROOT/local-ethernet-state gate that
VxAdmin uses for multi-station adjudication will be needed verbatim by
VxCentralScan networking. Move it to libs/networking as
isNetworkingEnabled(featureFlag) and have VxAdmin delegate to it.
