+++ libs/networking/src/index.ts
@@ -1,5 +1,6 @@
 /* istanbul ignore file */
 export { AvahiService, hasOnlineInterface } from './avahi.js';
 export type { AvahiDiscoveredService } from './avahi.js';
+export { isNetworkingEnabled } from './config.js';
 export { intermediateScript } from './intermediate_scripts.js';
 export { isValidIpv4Address } from './utils.js';
